==> jupyter/1. crawling phar data.ipynb <==
from selenium import webdriver
import pandas as pd
import time
from tqdm import tqdm
import re
import numpy as np

driver=webdriver.Chrome('/Users/gunwook/Desktop/chromedriver')

driver.get('http://find.pharm114.or.kr/find')

list_daeduck = []
for _ in tqdm(range(0,15)):
    bb = driver.find_elements_by_css_selector('#searchForm > div > div.date > span > img')[0]
    driver.execute_script("arguments[0].click();", bb)
    cc = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > a.ui-datepicker-next.ui-corner-all > span')[0]
    driver.execute_script("arguments[0].click();", cc)
    month = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > div')[0].text
    a = np.random.choice([2,3,4],1)[0]
    
    for j in range(1,8):
        week_a = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0})'.format(a))[0].text.split(' ')
        week = week_a[j-1]
        e = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0}) > td:nth-child({1})'.format(a,j))[0]
        driver.execute_script("arguments[0].click();", e)
        f = driver.find_element_by_css_selector('#searchBtn')
        driver.execute_script("arguments[0].click();", f)
        time.sleep(0.5)
        name = driver.find_elements_by_css_selector('#resultArea > tr > td.first_of_type.view')
        worktime = driver.find_elements_by_css_selector('#resultArea > tr > td.last_of_type2')
        location = driver.find_elements_by_css_selector('#resultArea > tr > td:nth-child(2)')
        ccc = driver.find_element_by_css_selector('#searchForm > div > div.date > span > img')
        driver.execute_script("arguments[0].click();", ccc)
        print(week)
        for b,c,d in zip(name,worktime,location):
            if '\n' in d.text:
                d = d.text.split('\n')[0]
                list_daeduck.append([b.text,d,c.text,subarea,month + ' ' + week + '일'])
            else:
                list_daeduck.append([b.text,d.text,c.text,subarea,month + ' ' + week + '일'])

대덕 = pd.DataFrame(list_daeduck, columns = ['약국명','주소','운영시간','지역','날짜'])

list_donggu = []
for _ in tqdm(range(0,15)):
    bb = driver.find_elements_by_css_selector('#searchForm > div > div.date > span > img')[0]
    driver.execute_script("arguments[0].click();", bb)
    cc = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > a.ui-datepicker-next.ui-corner-all > span')[0]
    driver.execute_script("arguments[0].click();", cc)
    month = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > div')[0].text
    a = np.random.choice([2,3,4],1)[0]
    
    for j in range(1,8):
        week_a = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0})'.format(a))[0].text.split(' ')
        week = week_a[j-1]
        e = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0}) > td:nth-child({1})'.format(a,j))[0]
        driver.execute_script("arguments[0].click();", e)
        f = driver.find_element_by_css_selector('#searchBtn')
        driver.execute_script("arguments[0].click();", f)
        time.sleep(0.5)
        name = driver.find_elements_by_css_selector('#resultArea > tr > td.first_of_type.view')
        worktime = driver.find_elements_by_css_selector('#resultArea > tr > td.last_of_type2')
        location = driver.find_elements_by_css_selector('#resultArea > tr > td:nth-child(2)')
        ccc = driver.find_element_by_css_selector('#searchForm > div > div.date > span > img')
        driver.execute_script("arguments[0].click();", ccc)
        print(week)
        for b,c,d in zip(name,worktime,location):
            if '\n' in d.text:
                d = d.text.split('\n')[0]
                list_donggu.append([b.text,d,c.text,subarea,month + ' ' + week + '일'])
            else:
                list_donggu.append([b.text,d.text,c.text,subarea,month + ' ' + week + '일'])

동구 = pd.DataFrame(list_donggu, columns = ['약국명','주소','운영시간','지역','날짜'])

list_joonggu = []
for _ in tqdm(range(0,15)):
    bb = driver.find_elements_by_css_selector('#searchForm > div > div.date > span > img')[0]
    driver.execute_script("arguments[0].click();", bb)
    cc = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > a.ui-datepicker-next.ui-corner-all > span')[0]
    driver.execute_script("arguments[0].click();", cc)
    month = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > div')[0].text
    a = np.random.choice([2,3,4],1)[0]
    
    for j in range(1,8):
        week_a = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0})'.format(a))[0].text.split(' ')
        week = week_a[j-1]
        e = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0}) > td:nth-child({1})'.format(a,j))[0]
        driver.execute_script("arguments[0].click();", e)
        f = driver.find_element_by_css_selector('#searchBtn')
        driver.execute_script("arguments[0].click();", f)
        time.sleep(0.5)
        name = driver.find_elements_by_css_selector('#resultArea > tr > td.first_of_type.view')
        worktime = driver.find_elements_by_css_selector('#resultArea > tr > td.last_of_type2')
        location = driver.find_elements_by_css_selector('#resultArea > tr > td:nth-child(2)')
        ccc = driver.find_element_by_css_selector('#searchForm > div > div.date > span > img')
        driver.execute_script("arguments[0].click();", ccc)
        print(week)
        for b,c,d in zip(name,worktime,location):
            if '\n' in d.text:
                d = d.text.split('\n')[0]
                list_joonggu.append([b.text,d,c.text,subarea,month + ' ' + week + '일'])
            else:
                list_joonggu.append([b.text,d.text,c.text,subarea,month + ' ' + week + '일'])

중구 = pd.DataFrame(list_joonggu, columns = ['약국명','주소','운영시간','지역','날짜'])

list_seogu = []
for _ in tqdm(range(0,15)):
    bb = driver.find_elements_by_css_selector('#searchForm > div > div.date > span > img')[0]
    driver.execute_script("arguments[0].click();", bb)
    cc = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > a.ui-datepicker-next.ui-corner-all > span')[0]
    driver.execute_script("arguments[0].click();", cc)
    month = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > div')[0].text
    a = np.random.choice([2,3,4],1)[0]
    
    for j in range(1,8):
        week_a = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0})'.format(a))[0].text.split(' ')
        week = week_a[j-1]
        e = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0}) > td:nth-child({1})'.format(a,j))[0]
        driver.execute_script("arguments[0].click();", e)
        f = driver.find_element_by_css_selector('#searchBtn')
        driver.execute_script("arguments[0].click();", f)
        time.sleep(0.5)
        name = driver.find_elements_by_css_selector('#resultArea > tr > td.first_of_type.view')
        worktime = driver.find_elements_by_css_selector('#resultArea > tr > td.last_of_type2')
        location = driver.find_elements_by_css_selector('#resultArea > tr > td:nth-child(2)')
        ccc = driver.find_element_by_css_selector('#searchForm > div > div.date > span > img')
        driver.execute_script("arguments[0].click();", ccc)
        print(week)
        for b,c,d in zip(name,worktime,location):
            if '\n' in d.text:
                d = d.text.split('\n')[0]
                list_seogu.append([b.text,d,c.text,subarea,month + ' ' + week + '일'])
            else:
                list_seogu.append([b.text,d.text,c.text,subarea,month + ' ' + week + '일'])

서구 = pd.DataFrame(list_seogu, columns = ['약국명','주소','운영시간','지역','날짜'])

list_yoosung = []
for _ in tqdm(range(0,15)):
    bb = driver.find_elements_by_css_selector('#searchForm > div > div.date > span > img')[0]
    driver.execute_script("arguments[0].click();", bb)
    cc = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > a.ui-datepicker-next.ui-corner-all > span')[0]
    driver.execute_script("arguments[0].click();", cc)
    month = driver.find_elements_by_css_selector('#ui-datepicker-div > div.ui-datepicker-header.ui-widget-header.ui-helper-clearfix.ui-corner-all > div')[0].text
    a = np.random.choice([2,3,4],1)[0]
    
    for j in range(1,8):
        week_a = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0})'.format(a))[0].text.split(' ')
        week = week_a[j-1]
        e = driver.find_elements_by_css_selector('#ui-datepicker-div > table > tbody > tr:nth-child({0}) > td:nth-child({1})'.format(a,j))[0]
        driver.execute_script("arguments[0].click();", e)
        f = driver.find_element_by_css_selector('#searchBtn')
        driver.execute_script("arguments[0].click();", f)
        time.sleep(0.5)
        name = driver.find_elements_by_css_selector('#resultArea > tr > td.first_of_type.view')
        worktime = driver.find_elements_by_css_selector('#resultArea > tr > td.last_of_type2')
        location = driver.find_elements_by_css_selector('#resultArea > tr > td:nth-child(2)')
        ccc = driver.find_element_by_css_selector('#searchForm > div > div.date > span > img')
        driver.execute_script("arguments[0].click();", ccc)
        print(week)
        for b,c,d in zip(name,worktime,location):
            if '\n' in d.text:
                d = d.text.split('\n')[0]
                list_yoosung.append([b.text,d,c.text,subarea,month + ' ' + week + '일'])
            else:
                list_yoosung.append([b.text,d.text,c.text,subarea,month + ' ' + week + '일'])

유성구 = pd.DataFrame(list_yoosung, columns = ['약국명','주소','운영시간','지역','날짜'])

final = pd.concat([대덕, 동구, 중구, 서구, 유성구])

final.to_csv('약국 운영 정보.csv')
