==> jupyter/2. preprocess (population).ipynb <==
# Data Preprocessing
### Resident population & Float population

import numpy as np
import pandas as pd
import re
from tqdm import tqdm

float_people = pd.read_excel('data/float_people.xlsx')
resid_people = pd.read_excel('data/resid_people.xls')

### - Concatening resident population columns

resid_people.iloc[:, ]

def concat_col(df, n):
    return df.iloc[:, n] + df.iloc[:, n+1]
    

resid_people.columns

total = 0
for i in (range(7)):
    n = 2*i+1
    resid_people.iloc[:, n] = (concat_col(resid_people, n))
    
for i in (range(7,18)):
    n = 2*i+1
    resid_people.iloc[:, n] = (concat_col(resid_people, n))

for i in range(30,37):
    total += resid_people.iloc[:, i]
    resid_people.iloc[:, 30] = total

drop_index = list(range(2,15,2))+list(range(17,30,2))+list(range(31,37))
resid_people.drop(columns=list(resid_people.columns[drop_index]), inplace=True)

### - Process as a proper type

float_people.columns

float_col = list(float_people.columns[5:])
resid_col = list(resid_people.columns[1:])
len(float_col) == len(resid_col)

col_change = {}
for i in range(len(float_col)):
    col_change[resid_col[i]] = float_col[i]

resid_people.rename(columns=col_change, inplace=True)

resid_people.rename(columns={'Unnamed: 0': '시'}, inplace=True)

resid_people.insert(1, '행정동', 0)
resid_people.insert(1, '구', 0)

df = resid_people.copy()

drop_row = []
for i in range(len(df)):
    if re.search('.+구 .+동', df.iloc[i,0]) == None:
        drop_row.append(i)
    else:
        df.iloc[i,1] = re.findall('대전광역시 (.+구) .+동', df.iloc[i,0])[0]
        df.iloc[i,2] = re.findall('대전광역시 .+구 (.+동)', df.iloc[i,0])[0]
        df.iloc[i,0] = '대전광역시'
        
df.drop(index = drop_row, inplace=True)

df

df.to_excel('resid_people_revised.xls')

float_people.shape

df = pd.DataFrame(data=np.zeros([455040,9]), columns=['시', '구', '행정동', '년', '월', '시간대', '성별', '연령대', '유동인구'])

df['시'] = '대전광역시'

index = 0
age = ['10대 미만', '10대', '20대', '30대', '40대', '50대', '60대', '70대 이상']
for i in tqdm(range(len(float_people))):
        df.loc[index:index+15, '구'] = float_people.loc[i, '구']
        df.loc[index:index+15, '행정동'] = float_people.loc[i, '행정동']
        df.loc[index:index+15, '년'] = str(float_people.loc[i, '연월'])[:4]
        df.loc[index:index+15, '월'] = str(float_people.loc[i, '연월'])[4:]
        df.loc[index:index+15, '시간대'] = float_people.loc[i, '시간대']
        
        df.loc[index:index+7, '성별'] = '남'
        df.loc[index+8:index+15, '성별'] = '여'
        df.loc[index:index+7, '연령대'] = age
        df.loc[index+8:index+15, '연령대'] = age
        
        df.loc[index:index+7, '유동인구'] = float_people.iloc[i, 5:13].values
        df.loc[index+8:index+15, '유동인구'] = float_people.iloc[i, 13:].values
        index+=16

df.to_csv('float_people_revised.csv', encoding='utf8')

resid = pd.read_excel('resid_people_revised.xls', index_col=0)

resid.reset_index(inplace=True)

df2 = pd.DataFrame(data = np.zeros([1264, 6]), columns=['시', '구', '행정동', '성별', '연령대', '거주인구'])
index = 0
age = ['10대 미만', '10대', '20대', '30대', '40대', '50대', '60대', '70대 이상']
df2['시'] = '대전광역시'
for i in tqdm(range(len(resid))):
        df2.loc[index:index+15, '구'] = resid.loc[i, '구']
        df2.loc[index:index+15, '행정동'] = resid.loc[i, '행정동']

        df2.loc[index:index+7, '성별'] = '남'
        df2.loc[index+8:index+15, '성별'] = '여'
        df2.loc[index:index+7, '연령대'] = age
        df2.loc[index+8:index+15, '연령대'] = age

        df2.loc[index:index+7, '거주인구'] = resid.iloc[i, 4:12].values
        df2.loc[index+8:index+15, '거주인구'] = resid.iloc[i, 12:].values
        index+=16

df2

df2.to_csv('resid_people_revised.csv')
