==> jupyter/4. caculate weakness index.ipynb <==
# 기본 설정

# colab 용

# Google Drive Mount

from google.colab import auth
auth.authenticate_user()

from google.colab import drive
drive.mount('/content/gdrive')

cd gdrive/My Drive/스스로공부하는능동적이고자율적인혀누

# 로컬용

import pandas as pd
import numpy as np
from tqdm import tqdm
pd.options.display.max_rows = 100
pd.options.display.max_columns = 100

# 지도시각화
import folium
from folium import plugins
import html
import json
from pandas.io.json import json_normalize
import os
import requests
import ipywidgets
from IPython.display import Image
from ipywidgets import interact
pd.set_option("display.max_rows", 220)
pd.set_option("display.max_columns", 220)

# ver 3.0로 하기

node = pd.read_csv('node v3.0.csv')
phar = pd.read_csv('약국정보_운영시간 ver2.0.csv')

## 거리 구하는 함수
import numbers
import math

class GeoUtil:
    """
    Geographical Utils
    """
    @staticmethod
    def degree2radius(degree):
        return degree * (math.pi/180)
    
    @staticmethod
    def get_harversion_distance(x1, y1, x2, y2, round_decimal_digits=5):
        """
        경위도 (x1,y1)과 (x2,y2) 점의 거리를 반환
        Harversion Formula 이용하여 2개의 경위도간 거래를 구함(단위:Km)
        """
        if x1 is None or y1 is None or x2 is None or y2 is None:
            return None
        assert isinstance(x1, numbers.Number) and -180 <= x1 and x1 <= 180
        assert isinstance(y1, numbers.Number) and  -90 <= y1 and y1 <=  90
        assert isinstance(x2, numbers.Number) and -180 <= x2 and x2 <= 180
        assert isinstance(y2, numbers.Number) and  -90 <= y2 and y2 <=  90

        R = 6371 # 지구의 반경(단위: km)
        dLon = GeoUtil.degree2radius(x2-x1)    
        dLat = GeoUtil.degree2radius(y2-y1)

        a = math.sin(dLat/2) * math.sin(dLat/2) \
            + (math.cos(GeoUtil.degree2radius(y1)) \
              *math.cos(GeoUtil.degree2radius(y2)) \
              *math.sin(dLon/2) * math.sin(dLon/2))
        b = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
        return round(R * b, round_decimal_digits)

    @staticmethod
    def get_euclidean_distance(x1, y1, x2, y2, round_decimal_digits=5):        
        """
        유클리안 Formula 이용하여 (x1,y1)과 (x2,y2) 점의 거리를 반환
        """
        if x1 is None or y1 is None or x2 is None or y2 is None:
            return None
        assert isinstance(x1, numbers.Number) and -180 <= x1 and x1 <= 180
        assert isinstance(y1, numbers.Number) and  -90 <= y1 and y1 <=  90
        assert isinstance(x2, numbers.Number) and -180 <= x2 and x2 <= 180
        assert isinstance(y2, numbers.Number) and  -90 <= y2 and y2 <=  90

        dLon = abs(x2-x1) # 경도 차이
        if dLon >= 180:   # 반대편으로 갈 수 있는 경우
            dLon -= 360   # 반대편 각을 구한다
        dLat = y2-y1      # 위도 차이
        return round(math.sqrt(pow(dLon,2)+pow(dLat,2)),round_decimal_digits)

def jsonify(data):
    df = json_normalize(data, '근처약국', ['날짜','시간'])
    df = df[['날짜', '시간','node_id','행정동', 'near_phar']]
    return df

phar[['날짜', '요일']].drop_duplicates().sort_values(by=['날짜']).values[:27]   # 현우
phar[['날짜', '요일']].drop_duplicates().sort_values(by=['날짜']).values[27:54] # 지환
phar[['날짜', '요일']].drop_duplicates().sort_values(by=['날짜']).values[54:81] # 건욱
phar[['날짜', '요일']].drop_duplicates().sort_values(by=['날짜']).values[81:]   # 세은

total_data = []
### 밑에 코드 주석처리에 자신의 할당량을 넣으시면 되뮤 ###
for target_date in tqdm(#phar[['날짜', '요일']].drop_duplicates().sort_values(by=['날짜'])['날짜'].values[]):
    for target_time in tqdm(range(6, 24)):
        df_temp_phar = phar[phar['날짜']== target_date]
        df_temp_phar = df_temp_phar[df_temp_phar[str(target_time)] == 1]  # 해당 날짜와 시간에 근무하는 약국 찾기

        list_node = list(node['node_id'].unique())
        list_phar_add = list(df_temp_phar['주소'].unique())
        list_near_phar = []
        for eachNode in list_node:
            cri_node_lon = node[node['node_id']==eachNode]['node_lon'].values[0]
            cri_node_lat = node[node['node_id']==eachNode]['node_lat'].values[0]
            list_target = list_node.copy()
            list_target.remove(eachNode)

            # node와 약국 사이의 거리 구하기
            list_collect_distance = []
            dict_num_of_phar = {}
            for phar_add in list_phar_add:
                target_lon = df_temp_phar[df_temp_phar['주소']==phar_add]['Longitude'].values[0]
                target_lat = df_temp_phar[df_temp_phar['주소']==phar_add]['Latitude'].values[0]
                list_collect_distance.append(GeoUtil.get_harversion_distance(cri_node_lon, cri_node_lat, target_lon, target_lat))
            num_of_phar = len([dist for dist in list_collect_distance if dist <= 0.5])
            dict_num_of_phar['node_id'] = eachNode
            dict_num_of_phar['행정동'] = node[node['node_id']==eachNode]['행정동'].values[0]
            dict_num_of_phar['near_phar'] = num_of_phar
            list_near_phar.append(dict_num_of_phar)

        dict_result = {}
        dict_result['날짜'] = target_date
        dict_result['시간'] = target_time
        dict_result['근처약국'] = list_near_phar      

    total_data.append(dict_result)

df = jsonify(total_data)

df.to_csv('몇 번부터 몇 번까지 했는지 적어줭~.csv', index=False)
# 예) '54~81.csv'
