==> jupyter/5. clustering.ipynb <==
# Data Analysis : Clustering

### - Make caculated columns

import pandas as pd
import numpy as np

float_ppl = pd.read_csv('float_people_revised.csv', index_col=0)
resid_ppl = pd.read_csv('resid_people_revised.csv', index_col=0)

float_ppl

df = pd.merge(float_ppl, resid_ppl, how='outer')

df.replace(' - ', 0, inplace=True)

df['유동인구'] = df['유동인구'].astype('int')

df_12 = df[df['시간대']==12]

df_fin = df_12.groupby('행정동')['유동인구', '거주인구'].sum()

df_fin['유동-거주인구비율_12시'] = (df_fin['유동인구'] / df_fin['거주인구']) * 5760

df_fin.to_csv('유동-거주인구비율_12시.csv')

### - Clustering

from sklearn.cluster import KMeans
from sklearn.datasets import make_blobs
import matplotlib.pylab as plt
import pandas as pd
import numpy as np
import seaborn as sns

df_cluster = pd.read_csv('clustering_v3.csv')

drop_col = ['유동인구', '거주인구', '10대 미만 합', '전체 합', 
            '10대', '10대 미만', '20대', '30대', '40대', '50대', '60대', '70대 이상'
           ,'유동-거주인구비율_12시', '속한 노드수']
df_cluster = df_cluster.drop(columns=drop_col)

df_cluster

X = df_cluster.drop(columns=['행정동'])
y = df_cluster['행정동']

kmeans = KMeans(n_clusters=4)
kmeans.fit(X)
y_pred = kmeans.fit_predict(X)

df_cluster['result'] = y_pred
df_cluster

df_cluster[(df_cluster['행정동']=='신성동') | (df_cluster['행정동']=='관평동') | (df_cluster['행정동']=='전민동')]

- GMM

from sklearn.mixture import GaussianMixture
gmm = GaussianMixture(n_components = , n_init = 10, random_state=40)
gmm.fit(X)

df_cluster['result'] = gmm.predict(X)

df_cluster[(df_cluster['행정동']=='신성동') | (df_cluster['행정동']=='관평동') | (df_cluster['행정동']=='전민동')]

df_cluster.groupby('result').count()

df_cluster[df_cluster['result']==3]

df_cluster.to_csv('clustering_gmm_v2.csv')
